+++ src/test/java/biz/gabrys/maven/plugins/css/splitter/steadystate/converter/FontFaceRuleConverterTest.java
@@ -1,6 +1,8 @@
 package biz.gabrys.maven.plugins.css.splitter.steadystate.converter;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 import java.util.List;
 
@@ -8,7 +10,6 @@ import org.junit.Test;
 
 import com.steadystate.css.dom.CSSFontFaceRuleImpl;
 import com.steadystate.css.dom.CSSStyleDeclarationImpl;
-import com.steadystate.css.dom.CSSValueImpl;
 import com.steadystate.css.dom.Property;
 
 import biz.gabrys.maven.plugins.css.splitter.css.type.StyleProperty;
@@ -28,25 +29,29 @@ public final class FontFaceRuleConverterTest {
 
     @Test
     public void convert() {
-        final FontFaceRuleConverter converter = new FontFaceRuleConverter();
+        final StylePropertyConverter stylePropertyConverter = mock(StylePropertyConverter.class);
+        final FontFaceRuleConverter converter = new FontFaceRuleConverter(stylePropertyConverter);
+
         final CSSFontFaceRuleImpl rule = new CSSFontFaceRuleImpl();
         final CSSStyleDeclarationImpl style = new CSSStyleDeclarationImpl(rule);
-        final Property property = new Property();
-        property.setName("name");
-        final CSSValueImpl value = new CSSValueImpl();
-        value.setCssText("value");
-        property.setValue(value);
-        style.addProperty(property);
+        final Property property1 = mock(Property.class);
+        style.addProperty(property1);
+        final Property property2 = mock(Property.class);
+        style.addProperty(property2);
         rule.setStyle(style);
 
+        final StyleProperty styleProperty1 = mock(StyleProperty.class);
+        when(stylePropertyConverter.convert(property1)).thenReturn(styleProperty1);
+        final StyleProperty styleProperty2 = mock(StyleProperty.class);
+        when(stylePropertyConverter.convert(property2)).thenReturn(styleProperty2);
+
         final StyleRule converted = converter.convert(rule);
+
         assertThat(converted).isNotNull();
         assertThat(converted.getSelectors()).containsExactly("@font-face");
         final List<StyleProperty> properties = converted.getProperties();
-        assertThat(properties).hasSize(1);
-        final StyleProperty styleProperty = properties.get(0);
-        assertThat(styleProperty).isNotNull();
-        assertThat(styleProperty.getName()).isEqualTo(property.getName());
-        assertThat(styleProperty.getValue()).isEqualTo(value.getCssText());
+        assertThat(properties).hasSize(2);
+        assertThat(properties.get(0)).isSameAs(styleProperty1);
+        assertThat(properties.get(1)).isSameAs(styleProperty2);
     }
 }