+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Features:
 * Set compatibility with Java 6+
 * Set compatibility with Maven 3+
 
+Bugs:
+* Fixed missing `!important` ([#23](https://github.com/gabrysbiz/css-splitter-maven-plugin/issues/23))
+
 Dependencies:
 * Upgraded [Maven Plugin Utils](http://maven-plugin-utils.projects.gabrys.biz/) from [1.4.1](http://maven-plugin-utils.projects.gabrys.biz/1.4.1/) to [2.0.0](http://maven-plugin-utils.projects.gabrys.biz/2.0.0/)